infrastructure/haystack/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
pgvector>=0.3.0
psycopg2-binary>=2.9.0
openai>=1.50.0
python-multipart>=0.0.12
